Implement rabbit, help and lots of things

+++ Guetta/Services/SocketClientEventsService.cs
@@ -56,6 +56,10 @@ namespace Guetta.Services
         private Task OnMessageReceived(DiscordClient sender, MessageCreateEventArgs messageCreateEventArgs)
         {
             var message = messageCreateEventArgs.Message;
+
+            if (message.Author.IsBot)
+                return Task.CompletedTask;
+            
             if (message.Content.StartsWith(CommandOptions.Value.Prefix))
             {
                 _ = message.DeleteMessageAfter(TimeSpan.FromSeconds(10));
